apps/object-detection/_old/object-counter/server-yolo/src/requirements.txt
numpy
fastapi
uvicorn
torch  
pydantic
opencv-python-headless  
Pillow
huggingface_hub
ultralytics
shutilwhich  
